requirements.txt
Flask==3.0.0
python-docx==1.1.0
markdown2==2.4.12
Pillow==10.2.0
html5lib==1.1
BeautifulSoup4==4.12.3
gunicorn==21.2.0
requests==2.31.0